Move Context & context/store => api/context & api/context/store
