+++ aci/backend.go
@@ -34,8 +34,8 @@ import (
 	"github.com/docker/compose-cli/api/volumes"
 
 	"github.com/docker/compose-cli/api/cloud"
-	apicontext "github.com/docker/compose-cli/context"
-	"github.com/docker/compose-cli/context/store"
+	apicontext "github.com/docker/compose-cli/api/context"
+	"github.com/docker/compose-cli/api/context/store"
 )
 
 const (
