+++ aci/context.go
@@ -27,7 +27,7 @@ import (
 	"github.com/hashicorp/go-uuid"
 	"github.com/pkg/errors"
 
-	"github.com/docker/compose-cli/context/store"
+	"github.com/docker/compose-cli/api/context/store"
 	"github.com/docker/compose-cli/errdefs"
 	"github.com/docker/compose-cli/prompt"
 )
