+++ aci/login/storagelogin.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/pkg/errors"
 
-	"github.com/docker/compose-cli/context/store"
+	"github.com/docker/compose-cli/api/context/store"
 )
 
 // StorageLogin helper for Azure Storage Login
