+++ cli/cmd/compose/up.go
@@ -24,8 +24,8 @@ import (
 
 	"github.com/docker/compose-cli/api/client"
 	"github.com/docker/compose-cli/api/compose"
+	"github.com/docker/compose-cli/api/context/store"
 	"github.com/docker/compose-cli/api/progress"
-	"github.com/docker/compose-cli/context/store"
 	"github.com/docker/compose-cli/formatter"
 
 	"github.com/compose-spec/compose-go/cli"
