+++ cli/cmd/context/create.go
@@ -23,8 +23,8 @@ import (
 
 	"github.com/spf13/cobra"
 
+	"github.com/docker/compose-cli/api/context/store"
 	"github.com/docker/compose-cli/cli/mobycli"
-	"github.com/docker/compose-cli/context/store"
 )
 
 type descriptionCreateOpts struct {
