+++ cli/cmd/context/create_aci.go
@@ -24,7 +24,7 @@ import (
 
 	"github.com/docker/compose-cli/aci"
 	"github.com/docker/compose-cli/api/client"
-	"github.com/docker/compose-cli/context/store"
+	"github.com/docker/compose-cli/api/context/store"
 	"github.com/docker/compose-cli/errdefs"
 )
