+++ cli/cmd/context/show.go
@@ -22,8 +22,8 @@ import (
 
 	"github.com/spf13/cobra"
 
-	apicontext "github.com/docker/compose-cli/context"
-	"github.com/docker/compose-cli/context/store"
+	apicontext "github.com/docker/compose-cli/api/context"
+	"github.com/docker/compose-cli/api/context/store"
 )
 
 func showCommand() *cobra.Command {
