+++ cli/mobycli/exec_test.go
@@ -21,7 +21,7 @@ import (
 
 	"gotest.tools/v3/assert"
 
-	"github.com/docker/compose-cli/context/store"
+	"github.com/docker/compose-cli/api/context/store"
 )
 
 func TestDelegateContextTypeToMoby(t *testing.T) {
