+++ cli/server/proxy/contexts_test.go
@@ -23,8 +23,8 @@ import (
 
 	"github.com/google/go-cmp/cmp/cmpopts"
 
+	"github.com/docker/compose-cli/api/context/store"
 	contextsv1 "github.com/docker/compose-cli/cli/server/protos/contexts/v1"
-	"github.com/docker/compose-cli/context/store"
 )
 
 func TestConvertContext(t *testing.T) {
