+++ docs/architecture.md
@@ -23,6 +23,6 @@ What follows is a list of useful links to help navigate the code:
   * An example backend can be found in [`example/`](../example)
 * The API is defined by protobufs that can be found in [`protos/`](../protos)
 * The API server is in [`server/`](../server)
-* The context management and interface can be found in [`context/`](../context)
+* The context management and interface can be found in [`context/`](../api/context)
 * The Node SDK is autogenerated (except for default endpoints managed by Docker Desktop), and can be found in
   [`docker/node-sdk`](https://github.com/docker/node-sdk)
