+++ tests/aci-e2e/e2e-aci_test.go
@@ -51,8 +51,8 @@ import (
 	"github.com/docker/compose-cli/aci/convert"
 	"github.com/docker/compose-cli/aci/login"
 	"github.com/docker/compose-cli/api/containers"
+	"github.com/docker/compose-cli/api/context/store"
 	"github.com/docker/compose-cli/cli/cmd"
-	"github.com/docker/compose-cli/context/store"
 	"github.com/docker/compose-cli/errdefs"
 	. "github.com/docker/compose-cli/tests/framework"
 )
